moves multer products from routes to middlewares folder
