+++ src/middlewares/multer-products.js
@@ -1,0 +1,17 @@
+const path = require("path");
+
+const multer = require("multer");
+
+const storage = multer.diskStorage({
+  destination: path.join(__dirname, "../../public/images/products"),
+  filename: function (req, file, cb) {
+    cb(
+      null,
+      file.fieldname + "-" + Date.now() + path.extname(file.originalname)
+    );
+  },
+});
+
+const upload = multer({ storage: storage });
+
+module.exports = upload;
